requirements.txt
furl==2.1.3
psutil==5.9.0
python-dotenv==0.19.2
requests==2.26.0
setuptools==58.1.0
validators==0.18.2
